backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
openai>=1.40,<2.0
python-dotenv>=1.0,<2.0
pydantic>=2.7,<3.0
